# openapi.yaml
openapi: 3.0.3
info:
  title: AIPAM API
  version: "2.0.0"
  description: |
    Air-gapped single-VM API for PCAP upload, job orchestration, sensor findings, and report/artifact access.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API_TOKEN
  parameters:
    RequestIdHeader:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
      description: Optional correlation ID. If provided, echoed back. If absent, server generates.
    JobIdPath:
      name: jobId
      in: path
      required: true
      schema: { type: string }
...
    HostIpPath:
      name: ip
      in: path
      required: true
      schema: { type: string }
      description: IP address string as stored/derived from job outputs.
    CursorQuery:
      name: cursor
      in: query
      required: false
      schema: { type: string }
      description: >
        Opaque pagination cursor. Stable only for the same filter/sort combination.
        If an old cursor is reused after filter/sort changes, returns 400 with code INVALID_CURSOR.
    LimitQuery:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    SortQuery:
      name: sort
      in: query
      required: false
      schema: { type: string }
      description: Sort field identifier for endpoint-specific list.
    OrderQuery:
      name: order
      in: query
      required: false
      schema:
        type: string
        enum: [asc, desc]
        default: desc

    # ---- Endpoint-specific query filters ----
...
    SearchQuery:
      name: q
      in: query
      required: false
      schema: { type: string }
      description: Free-text search / substring match.
    RoleFilter:
      name: role
      in: query
      required: false
      schema: { type: string, enum: [internal, external, unknown] }
    SeverityFilter:
      name: severity
      in: query
      required: false
      schema: { $ref: "#/components/schemas/Severity" }
...
    DirectionFilter:
      name: direction
      in: query
      required: false
      schema: { type: string, enum: [inbound, outbound, any], default: any }
    ProtoFilter:
      name: proto
      in: query
      required: false
      schema: { type: string, enum: [tcp, udp, icmp, any], default: any }
    DestIpFilter:
      name: dest_ip
      in: query
      required: false
      schema: { type: string }
    DestPortFilter:
      name: dest_port
      in: query
      required: false
      schema: { type: integer }
    ServiceFilter:
      name: service
      in: query
      required: false
      schema: { type: string }
    CommunityIdFilter:
      name: community_id
      in: query
      required: false
      schema: { type: string }
    StartTsFilter:
      name: start_ts
      in: query
      required: false
      schema: { type: string, format: date-time }
    EndTsFilter:
      name: end_ts
      in: query
      required: false
      schema: { type: string, format: date-time }
    RcodeFilter:
      name: rcode
      in: query
      required: false
      schema: { type: string }
    QtypeFilter:
      name: qtype
      in: query
      required: false
      schema: { type: string }
    SniFilter:
      name: sni
      in: query
      required: false
      schema: { type: string }
      description: SNI substring match.
    Ja3Filter:
      name: ja3
      in: query
      required: false
      schema: { type: string }
    IssuerFilter:
      name: issuer
      in: query
      required: false
      schema: { type: string }
      description: Certificate issuer substring match.
    CategoryFilter:
      name: category
      in: query
      required: false
      schema: { type: string }
    SidFilter:
      name: sid
      in: query
      required: false
      schema: { type: string }
    MimeFilter:
      name: mime
      in: query
      required: false
      schema: { type: string }
    Sha256Filter:
      name: sha256
      in: query
      required: false
      schema: { type: string }
    YaraRuleFilter:
      name: yara_rule
      in: query
      required: false
      schema: { type: string }
...
  responses:
    Unauthorized:
      description: Unauthorized (missing/invalid token)
      headers:
        X-Request-Id:
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Not found
      headers:
        X-Request-Id:
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    BadRequest:
      description: |
        Bad request (invalid input, cursor, limit, etc.).
        - If cursor is invalid or stale (e.g. reused after filter/sort change), returns code=INVALID_CURSOR.
        - If limit exceeds max (200), returns code=INVALID_LIMIT.
      headers:
        X-Request-Id:
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            invalid_cursor:
              summary: Stale or invalid cursor
              value: { code: "INVALID_CURSOR", message: "Cursor is invalid for current filter/sort combination.", details: { cursor: "abc123" } }
            invalid_limit:
              summary: Limit exceeds maximum
              value: { code: "INVALID_LIMIT", message: "Limit must be between 1 and 200.", details: { limit: 500, max: 200 } }
...
  schemas:
    SchemaVersion:
      type: string
      enum: ["1.0"]
      description: API schema version included in all responses.
    ErrorResponse:
      type: object
      required: [schema_version, code, message]
      properties:
        schema_version: { $ref: "#/components/schemas/SchemaVersion" }
        code: { type: string }
        message: { type: string }
        details:
          type: object
          additionalProperties: true
...
    Severity:
      type: string
      enum: [info, low, medium, high, critical]
...
    PageInfo:
      type: object
      required: [next_cursor, has_more]
      properties:
        next_cursor:
          type: string
          nullable: true
        has_more: { type: boolean }
...
    HostListItem:
      type: object
      required: [ip, role, conn_count, alert_count]
      properties:
        ip: { type: string }
        role:
          type: string
          enum: [internal, external, unknown]
        conn_count: { type: integer }
        bytes_sent: { type: integer, nullable: true }
        bytes_recv: { type: integer, nullable: true }
        alert_count: { type: integer }
        top_domains:
          type: array
          items: { type: string }
    HostDetail:
      allOf:
        - $ref: "#/components/schemas/HostListItem"
        - type: object
          properties:
            first_seen: { type: string, format: date-time, nullable: true }
            last_seen: { type: string, format: date-time, nullable: true }
            alerts_by_severity:
              type: object
              additionalProperties:
                type: integer
            top_services:
              type: array
              items: { type: string }
            dns_summary:
              type: object
              properties:
                query_count: { type: integer, nullable: true }
                top_qnames:
                  type: array
                  items: { type: string }
            tls_summary:
              type: object
              properties:
                session_count: { type: integer, nullable: true }
                top_sni:
                  type: array
                  items: { type: string }
                top_ja3:
                  type: array
                  items: { type: string }
    HostListResponse:
      type: object
      required: [schema_version, items, page]
      properties:
        schema_version: { $ref: "#/components/schemas/SchemaVersion" }
        items:
          type: array
          items: { $ref: "#/components/schemas/HostListItem" }
        page: { $ref: "#/components/schemas/PageInfo" }
    HostGetResponse:
      type: object
      required: [schema_version, host]
      properties:
        schema_version: { $ref: "#/components/schemas/SchemaVersion" }
        host: { $ref: "#/components/schemas/HostDetail" }
    ConnectionItem:
      type: object
      required: [connection_id, ts, src_ip, dest_ip, proto]
      properties:
        connection_id: { type: string }
        community_id: { type: string, nullable: true }
        ts: { type: string, format: date-time }
        src_ip: { type: string }
        src_port: { type: integer, nullable: true }
        dest_ip: { type: string }
        dest_port: { type: integer, nullable: true }
        proto: { type: string }
        duration_seconds: { type: number, nullable: true }
        bytes_sent: { type: integer, nullable: true }
        bytes_recv: { type: integer, nullable: true }
        service: { type: string, nullable: true }
        alerts:
          type: array
          items: { type: string }
        iocs:
          type: array
          items: { type: string }
    ConnectionListResponse:
      type: object
      required: [schema_version, items, page]
      properties:
        schema_version: { $ref: "#/components/schemas/SchemaVersion" }
        items:
          type: array
          items: { $ref: "#/components/schemas/ConnectionItem" }
        page: { $ref: "#/components/schemas/PageInfo" }
    DnsQueryItem:
      type: object
      required: [dns_id, ts, src_ip, query]
      properties:
        dns_id: { type: string }
        ts: { type: string, format: date-time }
        src_ip: { type: string }
        query: { type: string }
        qtype: { type: string, nullable: true }
        answers:
          type: array
          items: { type: string }
        rcode: { type: string, nullable: true }
        ttl_seconds: { type: integer, nullable: true }
        dest_ip: { type: string, nullable: true }
        community_id: { type: string, nullable: true }
        iocs:
          type: array
          items: { type: string }
        related_community_ids:
          type: array
          items: { type: string }
    DnsQueryListResponse:
      type: object
      required: [schema_version, items, page]
      properties:
        schema_version: { $ref: "#/components/schemas/SchemaVersion" }
        items:
          type: array
          items: { $ref: "#/components/schemas/DnsQueryItem" }
        page: { $ref: "#/components/schemas/PageInfo" }
    TlsSessionItem:
      type: object
      required: [tls_id, ts, src_ip, dest_ip]
      properties:
        tls_id: { type: string }
        ts: { type: string, format: date-time }
        src_ip: { type: string }
        dest_ip: { type: string }
        dest_port: { type: integer, nullable: true }
        sni: { type: string, nullable: true }
        ja3: { type: string, nullable: true }
        ja3s: { type: string, nullable: true }
        alpn: { type: string, nullable: true }
        version: { type: string, nullable: true }
        cert_subject: { type: string, nullable: true }
        cert_issuer: { type: string, nullable: true }
        cert_fingerprint_sha1: { type: string, nullable: true }
        community_id: { type: string, nullable: true }
        iocs:
          type: array
          items: { type: string }
    TlsSessionListResponse:
      type: object
      required: [schema_version, items, page]
      properties:
        schema_version: { $ref: "#/components/schemas/SchemaVersion" }
        items:
          type: array
          items: { $ref: "#/components/schemas/TlsSessionItem" }
        page: { $ref: "#/components/schemas/PageInfo" }
    AlertItem:
      type: object
      required: [alert_id, ts, severity, signature]
      properties:
        alert_id: { type: string }
        ts: { type: string, format: date-time }
        severity: { $ref: "#/components/schemas/Severity" }
        engine: { type: string, nullable: true }
        signature: { type: string }
        category: { type: string, nullable: true }
        sid: { type: string, nullable: true }
        src_ip: { type: string, nullable: true }
        src_port: { type: integer, nullable: true }
        dest_ip: { type: string, nullable: true }
        dest_port: { type: integer, nullable: true }
        proto: { type: string, nullable: true }
        community_id: { type: string, nullable: true }
        refs:
          type: array
          items: { type: string }
        tags:
          type: array
          items: { type: string }
    AlertListResponse:
      type: object
      required: [schema_version, items, page]
      properties:
        schema_version: { $ref: "#/components/schemas/SchemaVersion" }
        items:
          type: array
          items: { $ref: "#/components/schemas/AlertItem" }
        page: { $ref: "#/components/schemas/PageInfo" }
    FileItem:
      type: object
      required: [file_id, sha256, size_bytes]
      properties:
        file_id: { type: string }
        ts: { type: string, format: date-time, nullable: true }
        sha256: { type: string }
        md5: { type: string, nullable: true }
        ssdeep: { type: string, nullable: true }
        size_bytes: { type: integer }
        mime: { type: string, nullable: true }
        entropy: { type: number, nullable: true }
        source: { type: string, nullable: true }
        extracted_path: { type: string, nullable: true }
        yara_matches:
          type: array
          items: { type: string }
        download_artifact_id: { type: string, nullable: true }
        related_community_ids:
          type: array
          items: { type: string }
    FileListResponse:
      type: object
      required: [schema_version, items, page]
      properties:
        schema_version: { $ref: "#/components/schemas/SchemaVersion" }
        items:
          type: array
          items: { $ref: "#/components/schemas/FileItem" }
        page: { $ref: "#/components/schemas/PageInfo" }
...
paths:
...
  /jobs/{jobId}/hosts:
    get:
      tags: [Hosts]
      summary: List hosts in a job
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/JobIdPath"
        - $ref: "#/components/parameters/CursorQuery"
        - $ref: "#/components/parameters/LimitQuery"
        - $ref: "#/components/parameters/SortQuery"
        - $ref: "#/components/parameters/OrderQuery"
        - $ref: "#/components/parameters/RoleFilter"
        - $ref: "#/components/parameters/SearchQuery"
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HostListResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /jobs/{jobId}/hosts/{ip}:
    get:
      tags: [Hosts]
      summary: Host detail
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/JobIdPath"
        - $ref: "#/components/parameters/HostIpPath"
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HostGetResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /jobs/{jobId}/hosts/{ip}/connections:
    get:
      tags: [Hosts]
      summary: Host connections
      description: |
        Returns connections where the host IP ({ip}) appears as either source or destination.
        Use `direction=outbound` to match only connections where the host is the source,
        or `direction=inbound` where the host is the destination. Default (`any`) matches both.
        Default sort: `ts desc`.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/JobIdPath"
        - $ref: "#/components/parameters/HostIpPath"
        - $ref: "#/components/parameters/CursorQuery"
        - $ref: "#/components/parameters/LimitQuery"
        - $ref: "#/components/parameters/SortQuery"
        - $ref: "#/components/parameters/OrderQuery"
        - $ref: "#/components/parameters/DirectionFilter"
        - $ref: "#/components/parameters/ProtoFilter"
        - $ref: "#/components/parameters/DestIpFilter"
        - $ref: "#/components/parameters/DestPortFilter"
        - $ref: "#/components/parameters/ServiceFilter"
        - $ref: "#/components/parameters/CommunityIdFilter"
        - $ref: "#/components/parameters/StartTsFilter"
        - $ref: "#/components/parameters/EndTsFilter"
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConnectionListResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /jobs/{jobId}/hosts/{ip}/dns:
    get:
      tags: [Hosts]
      summary: Host DNS activity
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/JobIdPath"
        - $ref: "#/components/parameters/HostIpPath"
        - $ref: "#/components/parameters/CursorQuery"
        - $ref: "#/components/parameters/LimitQuery"
        - $ref: "#/components/parameters/SortQuery"
        - $ref: "#/components/parameters/OrderQuery"
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/RcodeFilter"
        - $ref: "#/components/parameters/QtypeFilter"
        - $ref: "#/components/parameters/CommunityIdFilter"
        - $ref: "#/components/parameters/StartTsFilter"
        - $ref: "#/components/parameters/EndTsFilter"
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DnsQueryListResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /jobs/{jobId}/hosts/{ip}/tls:
    get:
      tags: [Hosts]
      summary: Host TLS sessions
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/JobIdPath"
        - $ref: "#/components/parameters/HostIpPath"
        - $ref: "#/components/parameters/CursorQuery"
        - $ref: "#/components/parameters/LimitQuery"
        - $ref: "#/components/parameters/SortQuery"
        - $ref: "#/components/parameters/OrderQuery"
        - $ref: "#/components/parameters/SniFilter"
        - $ref: "#/components/parameters/Ja3Filter"
        - $ref: "#/components/parameters/IssuerFilter"
        - $ref: "#/components/parameters/CommunityIdFilter"
        - $ref: "#/components/parameters/StartTsFilter"
        - $ref: "#/components/parameters/EndTsFilter"
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TlsSessionListResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /jobs/{jobId}/hosts/{ip}/alerts:
    get:
      tags: [Hosts]
      summary: Alerts associated with host
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/JobIdPath"
        - $ref: "#/components/parameters/HostIpPath"
        - $ref: "#/components/parameters/CursorQuery"
        - $ref: "#/components/parameters/LimitQuery"
        - $ref: "#/components/parameters/SortQuery"
        - $ref: "#/components/parameters/OrderQuery"
        - $ref: "#/components/parameters/SeverityFilter"
        - $ref: "#/components/parameters/CategoryFilter"
        - $ref: "#/components/parameters/SidFilter"
        - $ref: "#/components/parameters/CommunityIdFilter"
        - $ref: "#/components/parameters/StartTsFilter"
        - $ref: "#/components/parameters/EndTsFilter"
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AlertListResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /jobs/{jobId}/hosts/{ip}/files:
    get:
      tags: [Hosts]
      summary: Files associated with host
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/JobIdPath"
        - $ref: "#/components/parameters/HostIpPath"
        - $ref: "#/components/parameters/CursorQuery"
        - $ref: "#/components/parameters/LimitQuery"
        - $ref: "#/components/parameters/SortQuery"
        - $ref: "#/components/parameters/OrderQuery"
        - $ref: "#/components/parameters/MimeFilter"
        - $ref: "#/components/parameters/Sha256Filter"
        - $ref: "#/components/parameters/YaraRuleFilter"
        - $ref: "#/components/parameters/CommunityIdFilter"
        - $ref: "#/components/parameters/StartTsFilter"
        - $ref: "#/components/parameters/EndTsFilter"
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FileListResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...